+++ cli/workspacelist.go
@@ -20,7 +20,11 @@ func workspaceList() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			workspaces, err := client.WorkspacesByUser(cmd.Context(), codersdk.Me)
+			organization, err := currentOrganization(cmd, client)
+			if err != nil {
+				return err
+			}
+			workspaces, err := client.WorkspacesByOwner(cmd.Context(), organization.ID, codersdk.Me)
 			if err != nil {
 				return err
 			}
