+++ cli/workspaceupdate.go
@@ -17,7 +17,11 @@ func workspaceUpdate() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			workspace, err := client.WorkspaceByName(cmd.Context(), codersdk.Me, args[0])
+			organization, err := currentOrganization(cmd, client)
+			if err != nil {
+				return err
+			}
+			workspace, err := client.WorkspaceByOwnerAndName(cmd.Context(), organization.ID, codersdk.Me, args[0])
 			if err != nil {
 				return err
 			}
